requirements.txt
# Core neuroimaging packages
nibabel>=5.0.0
nilearn>=0.10.0
nipype>=1.8.0

# Connectivity analysis
mne>=1.5.0
mne-connectivity>=0.5.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
niwidgets>=0.2.4

# Network analysis
networkx>=3.0
bct-python>=0.6.0  # Brain Connectivity Toolbox

# Image processing
scikit-image>=0.20.0
scikit-learn>=1.2.0

# Configuration
pyyaml>=6.0
configparser>=5.3.0

# Quality control
pybids>=0.15.0

# Statistical analysis
statsmodels>=0.14.0

# Progress bars
tqdm>=4.65.0
